Make all sub-menus detachable.
This includes the following type of menus: static, texture, plug-ins and xlinks.

Note for the texture menu:
There are two commands which result in the list of textures to be completely rebuilt, which will close all tearoff menus from the textures.  These commands are: "shaderlist.txt only" and "Directory list ..." .