Fix snapshots always being truncated to zero length .
The Map_Snapshot() function generated map names like "mymap.map.0", "mymap.map.1", etc. This name was then passed on to the saving method itself. However, the whole logic in which format to actually save the file depended on the extension being e.g. ".map". In these cases, the extension was ".<number>" and thus no appropriate saving mechanismus was found. The logic now creates files according to the scheme "mymap.0.map". This change was necessary anyway, because the editor cannot load a map file without the proper extension, too.